add sparse segtrees

+++ lib/basic_rbst.hpp
@@ -13,7 +13,7 @@ struct basic_rbst {
         node *l, *r;
         T val;
         bool rev;
-        u32 sz;
+        i32 sz;
 
         explicit node(const T &x)
             : l{nullptr}, r{nullptr}, val{x}, rev{false}, sz{1} {}
@@ -23,7 +23,7 @@ struct basic_rbst {
     };
 
     i32 size(node *t) const {
-        return t != nullptr ? static_cast<i32>(t->sz) : 0;
+        return t != nullptr ? t->sz : 0;
     }
 
     i32 n, pid;
