+++ lib/lazy_rbst.hpp
@@ -22,7 +22,7 @@ struct lazy_rbst {
         X val, sum, mus;
         A lz;
         bool rev;
-        u32 sz;
+        i32 sz;
 
         explicit node(const X &x)
             : l{nullptr}, r{nullptr}, val{x}, sum{x}, mus{x}, lz{MA::unit()}, rev{false}, sz{1} {}
@@ -32,7 +32,7 @@ struct lazy_rbst {
     };
 
     i32 size(node *t) const {
-        return t != nullptr ? static_cast<i32>(t->sz) : 0;
+        return t != nullptr ? t->sz : 0;
     }
 
     i32 n, pid;
