+++ lib/treap.hpp
@@ -17,7 +17,7 @@ struct treap {
         node *l, *r;
         X val, sum, mus;
         bool rev;
-        u32 sz;
+        i32 sz;
         u64 pr;
 
         explicit node(const X &x)
@@ -28,7 +28,7 @@ struct treap {
     };
 
     i32 size(node *t) const {
-        return t != nullptr ? static_cast<i32>(t->sz) : 0;
+        return t != nullptr ? t->sz : 0;
     }
 
     i32 n, pid;
